+++ bluetooth_low_energy_android/android/src/main/kotlin/dev/zeekr/bluetooth_low_energy_android/CentralManagerImpl.kt
@@ -227,6 +227,12 @@ class CentralManagerImpl(context: Context, binaryMessenger: BinaryMessenger) : B
     }
 
     @RequiresPermission(Manifest.permission.BLUETOOTH_CONNECT)
+    override fun requestConnectionPriority(addressArgs: String, priorityArgs: ConnectionPriorityArgs) {
+        val gatt = mGATTs[addressArgs] ?: throw IllegalArgumentException()
+        gatt.requestConnectionPriority(priorityArgs.toConnectionPriority())
+    }
+
+    @RequiresPermission(Manifest.permission.BLUETOOTH_CONNECT)
     override fun readRSSI(addressArgs: String, callback: (Result<Long>) -> Unit) {
         try {
             val gatt = mGATTs[addressArgs] ?: throw IllegalArgumentException()