+++ bluetooth_low_energy_windows/lib/src/central_manager_impl.dart
@@ -136,6 +136,11 @@ final class CentralManagerImpl
   }
 
   @override
+  Future<void> requestConnectionPriority(Peripheral peripheral, {required ConnectionPriority priority}) {
+    throw UnsupportedError('requestConnectionPriority is not supported on Windows.');
+  }
+
+  @override
   Future<int> getMaximumWriteLength(
     Peripheral peripheral, {
     required GATTCharacteristicWriteType type,